// apps/host/src/types/hyperliquid.ts
export interface PredictedFunding {
    fundingRate: string;
    nextFundingTime: number;
}